requirements.txt
fastapi
mangum
python-dotenv
openai 
fastembed
langchain
qdrant-client
uvicorn